CoreRamdiskPatcher: Ensure the SELinux context of /data/media/* is correct
Some ROMs do not have a line in /file_contexts that set the SELinux
context to u:object_r:media_rw_data_file:s0. When those ROMs call
restorecon, /data/media inherits the SELinux context of /data, which may
prevent ROMs with stricter SELinux policies from accessing the internal
storage.

This was not an issue before because secondary ROMs used /raw-data/media
as the source of the FUSE mountpoint and that location was not affected
by not affected by restorecon.