thirdparty: d-h.st -> dropbox
d-h.st seems to have a bug where files are randomly deleted when I try
to delete one specific file. It's quite frustrating to have to reupload
the tarballs all the time, so from now on, the prebuilt files will be
hosted on Dropbox.

Note: These are not rebuilds. The md5sums haven't changed.