libmbp: Improve gzip header detection in Loki'd boot images (fixes Kangakat)
This reduces the number of false positives when searching for the gzip
header by starting the search from (page size + kernel size), which is
the end of the kernel image (which always precedes the ramdisk).

Also, if multiple gzip headers are found, the ones with the 0x08 flag
will be preferred (meaning the header contains the original file name).
Most, if not all, Loki ramdisks are repacked and GNU gzip will set the
flag to 0x08.