mbtool: Give up on using ext4 images; go back to using directories
I thought using ext4 images would be a cool idea, but there are too many
factors that affect the resizing of those images. For example, we have:

- Kernel bugs on the Nexus 5 and Nexus 7 2013
- Incorrect time breaks e2fsck and resize2fs (eg. pulled out battery)
- system.new.dat has superblock specifying a size larger than the
  available free space on the destination partition
- etc.

It's just not worth the effort to make this work. We'll just go back to
using a plain old directory :)

(Of course, a temporary image is still used for installing Lollipop ROMs
that use the new block image update features.)