Android GUI: PatcherService: Fix horrible memory leak and switch to callbacks
PatchFileTask was previously holding an implicit reference to
PatcherService due to it being a non-static inner class. When we used
LocalBroadcastReceiver before, the issue was masked since it didn't
crash after onDestroy() was called. LeakCanary did complain, but didn't
log anything related to the leak. After modifying the service to use
callbacks, the issue was immediately noticeable as tasks created with
the previous instance of PatcherService did not use the callbacks.

The tasks are now static inner classes containing a WeakReference to the
service, which is updated in PatcherService.onCreate().