mbtool: roms.cpp: Scan mounted vold paths for MMC block devs if hard-coded SD card paths don't exist
Fixes in-app flashing when booted in a Marshmallow ROM. Vold now places
its fuse mountpoint in /storage/<ID>, where <ID> is the partition ID.
This makes it impossible to use hard-coded /storage/<blah> paths in
roms.cpp. This changes the code to look for /mnt/media_rw/<ID> mount
points associated with an MMC block dev. If the corresponding
/storage/<ID> path is a mount point, then return that path as the SD
card's mount point.

Eventually, we'll probably want a better way to deal with this. We could
potentially look for the sysfs path matching the SD card line in the
fstab file. Then, we look at all the uevent files to find the SD card
partition and grab the major and minor device numbers. Finally, we can
scan /proc/mounts and look for a block dev with those device numbers.