mbtool: Initial support for launching the boot UI
This commits makes the necessary changes to allow the boot UI to run and
communicate with mbtool. init will now mount the basic mount points
(procfs, sysfs, etc.) and the basic partitions (system, cache, data) and
then start up an instance of the mbtool daemon. The boot UI is then
launched and is responsible for controlling which ROM to boot. If the
boot UI exits or crashes in any way, then boot process continues. The
boot UI must talk to the daemon to switch to another ROM and reboot.

The boot UI is a zip file (and signature) containing file named 'exec'
and is located at/raw/cache/multiboot/bootui.zip. init will extract
'exec', make it executable, and launch it, passing the path to the zip
file as the first argument. The boot UI will then handle the extraction
of the theme files from the zip.

If the user decides to switch to another ROM, the boot UI will instruct
the daemon to switch ROMs and then it will write the ROM ID to
/raw/cache/multiboot/bootui/skip before rebooting. On the next boot,
init will read this file and if it's contents match the actual ROM ID,
then the boot menu will be skipped and the boot process will continue.
This avoids having the boot menu show up twice.