libmbutil: Fix remove-while-iterating bug in util::unmount_all()
The function was originally unmounting filesystems during a getmntent()
loop. This caused issues since the contents of /proc/mounts changed when
a filesystem was unmounted, but the file position remained the same.

This is now fixed by storing the matching mountpoints in a list and then
reverse-iterating over the list to unmount the filesystems. This also
reduces the number of retries we need to everything in a directory tree.