mbbootui: Fix null pointer dereferencing if daemon returns incomplete ROM data
When getting the list of installed ROMs from the daemon, the client did
not check if the fields (system path, build, version, etc.) were NULL
before accessing them. This would cause the Boot UI to crash upon
selection of the Switch ROM page. Under normal operation, this could
happen if, for any ROM:

* The build.prop file is not found or not readable
* The build.prop file is missing properties for the build and version
* The ROM uses a system image and thus, the build.prop file cannot be
  read