libmbp: Strip 'erase' command in system.transfer.list due to BLKDISCARD bugs on some devices
CM 13 and earlier contained an updater-script file that did not verify
the result of block_image_update(). If a BLKDISCARD ioctl call failed
before, it would simply be ignored. This changed with CM 14, which
properly checks the return value of block_image_update(). We'll work
around the issue by removing the 'erase' command in
system.transfer.list.