mbtool: Wait for SD card to be available
This is necessary as not every device detects the SD card early in the
boot process. With this change, /raw/extsd is no longer mounted unless
the ROM ID is an extsd slot. This is necessary because we can't
differentiate between the SD card not being inserted and the SD card not
being detected yet.