Drop support for encryption
The encryption support is extremely hackish and has no support for
hardware encryption or using keymaster. This makes it completely useless
for any device other than those made 3 or 4 years ago. And even then, it
only works with AOSP-based ROMs. As such, there's no reason to keep the
code around.