mbtool: Switch to libmbbootimg and patch boot image at flash time
Boot images will no longer be patched by libmbp. Instead, the required
binaries will be added to the zip and mbtool will handle the patching of
the ramdisk. This allows us to guarantee that the ramdisk will be
patched correctly, regardless of how it's stored in the original zip
file.

The only remaining code that uses mbp::BootImage and mbp::CpioFile is
the ramdisk update feature. That will be ported to libmbbootimg in
another commit.