cmake: Remove useless OBJECT targets
The way they were used prevented the use of modern CMake features, such
as INTERFACE properties. They would be added to the object targets,
instead of the library targets, which make them useless for the
dependees of the library target.

The object targets were originally used by tests that need to link to
internal symbols. These tests targets have been modified to simply link
to the static version of the libraries.

Signed-off-by: Andrew Gunnerson <chenxiaolong@cxl.epac.to>