android: Use shared libstdc++ for android-app target
This was an oversight when DualBootPatcher switched from having just
libmbp.so to having several different libraries. When linked against
static libstdc++, symbols that were meant to be globally unique got
duplicated in each library.

This was particularly problematic for the
_ZNSs4_Rep20_S_empty_rep_storageE symbol. It's a statically allocated
buffer used as the representation for empty strings. If an empty
std::string is passed from one library to another and the second
library performs an action that causes a reallocation, then the second
library will free() the _S_empty_rep_storage buffer and crash. The
normal `this != &_S_empty_rep()` guard doesn't work because each
library has its own _S_empty_rep_storage.

Signed-off-by: Andrew Gunnerson <chenxiaolong@cxl.epac.to>