libmbcommon: file: Fix possible segfault caused by moving into a mb::File reference
The move assignment operator was previously public and since references
are polymorphic, the following (valid) code could be written:

    PosixFile file("/dev/null", FileOpenMode::ReadOnly);
    File &ref = file;
    ref = MemoryFile("", 0);

The move assignment operator will happily move the pimpl pointer of the
MemoryFile to the PosixFile. When the PosixFile falls out of scope and
calls close(), it will cast the pimpl pointer to PosixFilePrivate * and
dereference it, leading to a crash.

The short term fix is to just make the move constructor and assignment
operator protected in mb::File. In the future, we may want to make
mb::File unmovable, like std::ostream.

Signed-off-by: Andrew Gunnerson <chenxiaolong@cxl.epac.to>