mbtool: packages.cpp: Parse packages.xml flags as signed integers
The change to mb::str_to_num() that prevented negative numbers from
being parsed as unsigned integers caused the packages.xml parser to
break on newer versions of Android. The Android framework stores the
packages.xml flags as signed 32-bit integers and (1 << 31) is a valid
flag, so the number in the XML file ends up being negative.

Signed-off-by: Andrew Gunnerson <chenxiaolong@cxl.epac.to>