Include alternative codename of Axon 7
Add alternative codename 'axon7' to support more recoveries, especially newer ones that report the Axon 7 as 'axon7' rather than the old 'ailsa_ii'. Reason is that at the moment dualboot patcher patched zips for the Axon 7 refuse to be flashed on newer (Oreo) recoveries because the device is called 'axon7' now.