docker: Ensure HOME is always set properly
Previously, when USER_ID was set to a user that already existed in the
image, the entrypoint script would set HOME and then run gosu. However,
gosu itself sets HOME to whatever is in /etc/passwd.

This adds a second script that gets run as the target user and just sets
HOME properly.

Signed-off-by: Andrew Gunnerson <andrewgunnerson@gmail.com>