libmbsign: Complete rewrite to use ed25519 signatures
libmbsign now reads and writes minisign-compatible ed25519 signatures
instead of using a custom RSA signature format. The secret and public
keys are also compatible with minisign.

For simplicity, only PureEdDSA is implemented. If signing and verifying
large files is needed in the future, HashEdDSA will be implemented.

This change also switches the crypto library dependency from
OpenSSL/BoringSSL to libsodium, which should help reduce the binary size
of mbtool.

Signed-off-by: Andrew Gunnerson <andrewgunnerson@gmail.com>