libmbcommon: MemoryFile: Fix double-free after truncate(0)
If truncate(0) is called, then realloc(ptr, 0) is called, which is
allowed to either return NULL or a special pointer that can be passed to
free(). With glibc, NULL is returned, which causes truncate(0) to fail
and m_data does not get updated properly. When the MemoryFile is
destructed, the already-freed m_data would be freed again.

Signed-off-by: Andrew Gunnerson <andrewgunnerson@gmail.com>